Moved sound types to separate file
-Deduped sound structs that are already identified elsewhere.
